Trace rows of x-pixels in parallel
The loop over x-pixlels is replaced with a parallel for_each. This means that a row of pixels are traced in a parallel, but each row is sequential. The current approach does all work first and updates the image afterwards, which means that the order of execution matters.

Work:  O(ns)*O(|world|). Tracing ns rays up to a depth of 50 and reduce to a final pixel color.
Depth: O(nx) size of a pixel-row.